refactored controllers and services, updated seeding scripts

+++ apps/engineering/src/modules/booking/service/booking.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { BookingModel } from 'apps/engineering/src/entities/booking.model';
+import { AllBookingResponseContract, BookingRequestContract } from '../dto/booking.contracts';
 
 @Injectable()
 export class BookingService {
@@ -10,19 +11,46 @@ export class BookingService {
     private readonly bookingRepository: Repository<BookingModel>,
   ) {}
 
-  async findAllBookings(): Promise<BookingModel[]> {
-    return await this.bookingRepository.find();
+  async findAllBookingsAndFormat(): Promise<AllBookingResponseContract> {
+    const bookings = await this.bookingRepository.find();
+    const formattedBookings = bookings.map((booking) => ({
+      id: booking.id,
+      user: booking.user,
+      parc: booking.parc,
+      bookingdate: booking.bookingdate,
+      comments: booking.comments,
+    }));
+
+    return { data: formattedBookings };
   }
 
   async getBookingById(id: string): Promise<BookingModel> {
-    return await this.bookingRepository.createQueryBuilder().select().where('id = :id', {id}).getOne();
+    const bookingById =  await this.bookingRepository.findOne({where: {id}});
+    if (!bookingById) {
+      throw new NotFoundException('Booking not found');
+    }
+    return bookingById;
   }
 
-  async newBooking(booking: BookingModel): Promise<BookingModel> {
-    return await this.bookingRepository.save(booking);
+  async newBooking(booking: BookingRequestContract): Promise<BookingModel> {
+    try {
+      const newBooking = this.bookingRepository.create(booking);
+      await this.bookingRepository.save(newBooking);
+      return newBooking;
+    } catch (error) {
+      throw new BadRequestException(error.message);
+    }
   }
 
   async removeBooking(id: string): Promise<void> {
-    await this.bookingRepository.delete(id);
+    try {
+      const booking = await this.getBookingById(id);
+      if (!booking) {
+        throw new NotFoundException('Booking not found');
+      }
+      await this.bookingRepository.delete(id);
+    } catch (error) {
+      throw new BadRequestException(error.message);
+    }
   }
 }
